% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors_table}
\alias{colors_table}
\title{A table with different colors combinations for the theme}
\format{
A data frame with 644 rows and 3 variables:
\describe{
  \item{color}{color set name}
  \item{base}{base color, for text and }
  \item{light}{a lighter color}
  \item{background}{color for the background of plots}
  \item{dark}{a darker color}
  ...
}
}
\usage{
colors_table
}
\description{
A table containing the name of the color combination, a base color,
a background color, a ligther and a darker one
}
\keyword{datasets}
