% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_mediocre.R
\name{palette_mediocre_d}
\alias{palette_mediocre_d}
\title{Mediocre color palette}
\usage{
palette_mediocre_d(pal = "autumn", second_pair = FALSE, ...)
}
\arguments{
\item{pal}{a string. Defines the color palette selected.
Can either be either "autumn", "rainbow", "green",
 "hotcold", "blackandwhite", or "coty"}

\item{second_pair}{if true sets an alternative pair of colors
for graphs with two colors}

\item{...}{Other arguments passed on to
\code{\link[ggplot2]{discrete_scale}},
\code{\link[ggplot2]{continuous_scale}},
or \code{\link[ggplot2]{binned_scale}} to control name, limits, breaks,
labels and so forth.}
}
\description{
The `mediocre` palettes provide alternative default colors for ggplots.
These functions enable to create such palettes.
The main color palettes have been defined using \url{https://coolors.co/}.
}
\examples{
palette_mediocre_d()

palette_mediocre_c(pal = "rainbow")

}
\seealso{
Other color scales: 
\code{\link{palette_mediocre_c}()}
}
\concept{color scales}
