% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_mediocre.R, R/scales_mediocre.R
\name{palette_mediocre_c}
\alias{palette_mediocre_c}
\alias{scale_mediocre_d}
\alias{scale_mediocre_c}
\title{Mediocre color scales}
\usage{
palette_mediocre_c(pal = "autumn", ...)

scale_mediocre_d(pal = "autumn", second_pair = FALSE, ...)

scale_mediocre_c(pal = "autumn", ...)
}
\arguments{
\item{pal}{a string. Defines the color palette selected.
Can either be "autumn" or "rainbow"}

\item{...}{Other arguments passed on to
\code{\link[ggplot2]{discrete_scale}},
\code{\link[ggplot2]{continuous_scale}},
or \code{\link[ggplot2]{binned_scale}} to control name, limits, breaks,
labels and so forth.}

\item{second_pair}{if true sets an alternative pair of colors
for graphs with two colors}
}
\description{
The `mediocre` scales provide alternative default colors for ggplots.
These functions enable to create such palettes.
The main color palettes have been defined using \url{https://coolors.co/}.
}
\examples{
ggplot(ggplot2::mpg, aes(x = cty, y = displ, fill = manufacturer)) +
  geom_col() +
  theme_mediocre() +
  scale_mediocre_d()

ggplot(ggplot2::mpg, aes(x = cty, y = displ, color = hwy)) +
  geom_point() +
  theme_mediocre() +
  scale_mediocre_c()

}
\seealso{
Other color scales: 
\code{\link{palette_mediocre_d}()}
}
\concept{color scales}
