% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_document_mediocre.R
\name{set_document_mediocre}
\alias{set_document_mediocre}
\title{Sets the mediocre theme and colors for the whole document.}
\usage{
set_document_mediocre()
}
\description{
Sets the mediocre theme and colors for the whole document.
}
\examples{
set_document_mediocre()

}
