% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mediocre.R
\name{theme_mediocre}
\alias{theme_mediocre}
\alias{theme_mediocre_map}
\title{Defines the ggplot mediocretheme with the default R colors.}
\usage{
theme_mediocre(
  pal = "autumn",
  gradient = NULL,
  background = FALSE,
  black_text = TRUE,
  base_size = if (.Platform$OS.type == "windows") 17 else 14,
  base_family = "Lato",
  base_line_size = base_size/50,
  base_rect_size = base_size/170
)

theme_mediocre_map(
  pal = "autumn",
  gradient = NULL,
  background = FALSE,
  black_text = FALSE,
  base_size = if (.Platform$OS.type == "windows") 17 else 14,
  base_family = "Lato",
  base_line_size = base_size/50,
  base_rect_size = base_size/170
)
}
\arguments{
\item{pal}{a string. Defines the color palette selected.
Can either be either "autumn", "rainbow", "green",
 "hotcold", "blackandwhite", "coty", "leo", "portal" or "pem"}

\item{gradient}{a string, either "left" or "right".
Turns divergent palettes into gradient ones by splitting it in two.
If equal to "right", the gradient palette is built with the right hand
side part of the divergent palette. If equal to "left", uses the left
hand side but flipping so that light values are on the left.}

\item{background}{If TRUE, a background is added to the plots}

\item{black_text}{If TRUE, the text will be in black, overriding the palette
text color}

\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family}

\item{base_line_size}{Base size for line elements}

\item{base_rect_size}{Base size for rect elements}
}
\description{
Defines the ggplot mediocretheme with the default R colors.
}
\examples{
ggplot(ggplot2::mpg, aes(x = cty, y = displ, fill = manufacturer)) +
  geom_col() +
  theme_mediocre() +
  labs(title = "A very nice title", subtitle = "A disappointing subtitle")

ggplot(ggplot2::mpg, aes(x = cty, y = displ, color = hwy)) +
  geom_point() +
  theme_mediocre() +
  labs(title = "A very nice title", subtitle = "A disappointing subtitle")

}
