% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mediocre_only.R
\name{theme_mediocre_only}
\alias{theme_mediocre_only}
\title{Defines the ggplot theme with the default R colors}
\usage{
theme_mediocre_only(color = "gray")
}
\description{
Defines the ggplot theme with the default R colors
}
\examples{
ggplot(data = ggplot2::mpg, aes(x = cty, y = displ, fill = manufacturer)) +
geom_col() +
theme_mediocre_only() +
labs(title = "A very nice title", subtitle = "A desapointing subtitle")

ggplot(data = ggplot2::mpg, aes(x = cty, y = displ, fill = manufacturer)) +
geom_point() +
theme_mediocre_only() +
labs(title = "A very nice title", subtitle = "A desapointing subtitle")

}
