% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unset_mediocre_all.R
\name{unset_mediocre_all}
\alias{unset_mediocre_all}
\title{Unsets the mediocre theme and colors, reverting to the ggplot2 defaults.}
\usage{
unset_mediocre_all()
}
\description{
Unsets the mediocre theme and colors, reverting to the ggplot2 defaults.
}
\examples{
unset_mediocre_all()

}
