% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringan_mediocre.R
\name{xaringan_mediocre}
\alias{xaringan_mediocre}
\title{Mediocre xanringan theme}
\usage{
xaringan_mediocre(
  pal = "autumn",
  theme = mediocrethemes::theme_mediocre,
  mediocre_plots = TRUE
)
}
\arguments{
\item{pal}{a string. Defines the color palette selected.
Can either be either "autumn", "rainbow", "green",
 "hotcold", "blackandwhite", "coty", "leo", "portal" or "pem"
 "hotcold", "blackandwhite", "coty", "leo", "portal" or "pem"}

\item{mediocre_plots}{if TRUE, sets the plots of the .Rmd document
to be made with mediocrethemes (calls \code{set_mediocre_all(pal)})}
}
\value{
This function adds a CSS file to your current working directory.
}
\description{
This function creates a CSS file \code{mediocre-xaringan.css} to build
\href{https://pkg.garrickadenbuie.com/xaringanthemer/}{\code{xaringanthemer}}
slides, based on mediocrethemes style, fonts and colors.
}
\details{
For more information on how to use this function,
please refer to the associated vignette.
}
\examples{
\dontrun{
xaringan_mediocre()

xaringan_mediocre(pal = "rainbow")
}
}
