% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors_table}
\alias{colors_table}
\title{A table with different colors combinations for the theme}
\format{
A data frame with 644 rows and 3 variables:
\describe{
  \item{color}{color set name}
  \item{base}{base color, for one color graphs}
  \item{text}{color for text and axes}
  \item{light}{a lighter color}
  \item{background}{color for the background of plots}
  \item{dark}{a darker color}
  \item{complementary}{a complementary color to the base color, for contrast}
  \item{vector}{a vector of colors used in the scale}
  ...
}
}
\usage{
colors_table
}
\description{
A table containing the name of the color combination, a base color,
a background color, a lighter and a darker one
}
\keyword{datasets}
