% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes_mediocre.R
\name{palette_mediocre_d}
\alias{palette_mediocre_d}
\alias{palette_mediocre_c}
\title{Mediocre color palette}
\usage{
palette_mediocre_d(pal = "autumn", second_pair = FALSE, gradient = NULL, ...)

palette_mediocre_c(pal = "autumn", gradient = NULL, ...)
}
\arguments{
\item{pal}{a string. Defines the color palette selected.
Can either be either "autumn", "rainbow", "green",
 "hotcold", "blackandwhite", "coty", "leo", "portal" or "pem"}

\item{second_pair}{if true sets an alternative pair of colors
for graphs with two colors}

\item{gradient}{a string, either "left" or "right".
Turns divergent palettes into gradient ones by splitting it in two.
If equal to "right", the gradient palette is built with the right hand
side part of the divergent palette. If equal to "left", uses the left
hand side but flipping so that light values are on the left.}

\item{...}{Other arguments passed on to
\code{\link[ggplot2]{discrete_scale}},
\code{\link[ggplot2]{continuous_scale}},
or \code{\link[ggplot2]{binned_scale}} to control name, limits, breaks,
labels and so forth.}
}
\description{
The `mediocre` palettes provide alternative default colors for ggplots.
These functions enable to create such palettes.
The main color palettes have been defined using \url{https://coolors.co/}.
}
\examples{
palette_mediocre_d()

palette_mediocre_c(pal = "rainbow")

}
\seealso{
Other color scales: 
\code{\link{scale_mediocre_d}()}

Other color scales: 
\code{\link{scale_mediocre_d}()}
}
\concept{color scales}
