% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_mediocre.R
\name{scale_mediocre_d}
\alias{scale_mediocre_d}
\alias{scale_mediocre_c}
\title{Mediocre color scales}
\usage{
scale_mediocre_d(...)

scale_mediocre_c(...)
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[ggplot2]{discrete_scale}},
\code{\link[ggplot2]{continuous_scale}}, or \code{\link[ggplot2]{binned_scale}} to control name, limits, breaks,
  labels and so forth.}
}
\description{
The `mediocre` scales provide alternative default colors for ggplots.
The main color scale has been chosen and created using \url{https://coolors.co/}.
}
\examples{
ggplot(ggplot2::mpg, aes(x = cty, y = displ, fill = manufacturer)) +
  geom_col() +
  theme_mediocre() +
  scale_mediocre_d()

ggplot(ggplot2::mpg, aes(x = cty, y = displ, color = hwy)) +
  geom_point() +
  theme_mediocre() +
  scale_mediocre_c()

}
\concept{colour scales}
