% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mediocre_all.R
\name{set_mediocre_all}
\alias{set_mediocre_all}
\title{Sets the mediocre theme and colors for the whole document.}
\usage{
set_mediocre_all(
  pal = "autumn",
  theme = "mediocre",
  gradient = NULL,
  background = FALSE,
  black_text = FALSE,
  second_pair = FALSE,
  base_size = NULL,
  base_family = NULL,
  base_line_size = NULL,
  base_rect_size = NULL
)
}
\arguments{
\item{pal}{a string. Defines the color palette selected.
Can either be either "autumn", "rainbow", "green",
 "hotcold", "blackandwhite", "coty", "leo", "portal" or "pem"}

\item{theme}{a string. Defines the theme used, either "mediocre" or "idid"}

\item{gradient}{a string, either "left" or "right".
Turns divergent palettes into gradient ones by splitting it in two.
If equal to "right", the gradient palette is built with the right hand
side part of the divergent palette. If equal to "left", uses the left
hand side but flipping so that light values are on the left.}

\item{background}{If TRUE, a background is added to the plots}

\item{black_text}{If TRUE, the text will be in black, overriding the palette
text color}

\item{second_pair}{if true sets an alternative pair of colors
for graphs with two colors}

\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family}

\item{base_line_size}{Base size for line elements}

\item{base_rect_size}{Base size for rect elements}
}
\description{
Sets the mediocre theme and colors for the whole document.
}
\examples{
set_mediocre_all()

}
