% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mediocre_all.R
\name{set_mediocre_all}
\alias{set_mediocre_all}
\title{Sets the mediocre theme and colors for the whole document.}
\usage{
set_mediocre_all()
}
\description{
Sets the mediocre theme and colors for the whole document.
}
\examples{
set_mediocre_all()

}
