% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mediocre_all.R
\name{set_mediocre_all}
\alias{set_mediocre_all}
\title{Sets the mediocre theme and colors for the whole document.}
\usage{
set_mediocre_all(pal = "autumn", background = FALSE, second_pair = FALSE)
}
\arguments{
\item{pal}{a string. Defines the color palette selected.
Can either be "autumn" or "rainbow"}

\item{background}{If TRUE, a background is added to the plots}

\item{second_pair}{if true sets an alternative pair of colors
for graphs with two colors}
}
\description{
Sets the mediocre theme and colors for the whole document.
}
\examples{
set_mediocre_all()

}
