% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mediocre.R
\name{theme_mediocre}
\alias{theme_mediocre}
\title{Defines the ggplot theme with the default R colors.}
\usage{
theme_mediocre(pal = "autumn", background = FALSE)
}
\arguments{
\item{pal}{a string. Defines the color palette selected.
Can either be "autumn" or "rainbow"}

\item{background}{If TRUE, a background is added to the plots}
}
\description{
Defines the ggplot theme with the default R colors.
}
\examples{
ggplot(ggplot2::mpg, aes(x = cty, y = displ, fill = manufacturer)) +
  geom_col() +
  theme_mediocre() +
  labs(title = "A very nice title", subtitle = "A disappointing subtitle")

ggplot(ggplot2::mpg, aes(x = cty, y = displ, color = hwy)) +
  geom_point() +
  theme_mediocre() +
  labs(title = "A very nice title", subtitle = "A disappointing subtitle")

}
